Break out build tool executable into its own target
